Add lifecycle smoke test ci plugin
This commit adds a plugin that is heavily based on the one used by the
AOT smoke tests. It uses the description of each lifecycle smoke tests
to generate a GitHub Actions workflow with the necessary jobs.

There is also an additional task that generates or update a status
document that list the smoke tests per generation with a link to the
code and workflow.

See gh-61